rename platform to system_info
