+++ lbrynet/lbrynet_daemon/Daemon.py
@@ -39,7 +39,8 @@ from lbrynet.lbrynet_daemon.Downloader import GetStream
 from lbrynet.lbrynet_daemon.Publisher import Publisher
 from lbrynet.lbrynet_daemon.ExchangeRateManager import ExchangeRateManager
 from lbrynet.lbrynet_daemon.auth.server import AuthJSONRPCServer
-from lbrynet.core import log_support, utils, Platform
+from lbrynet.core import log_support, utils
+from lbrynet.core import system_info
 from lbrynet.core.StreamDescriptor import StreamDescriptorIdentifier, download_sd_blob
 from lbrynet.core.StreamDescriptor import BlobStreamDescriptorReader
 from lbrynet.core.Session import Session
@@ -354,7 +355,7 @@ class Daemon(AuthJSONRPCServer):
 
     def _get_platform(self):
         if self.platform is None:
-            self.platform = Platform.get_platform()
+            self.platform = system_info.get_platform()
             self.platform["ui_version"] = self.lbry_ui_manager.loaded_git_version
         return self.platform
